Finished refactoring of task list based config prep
