+++ src/swell/deployment/bin/swell_prepare_config.py
@@ -20,7 +20,7 @@ from swell.utilities.welcome_message import write_welcome_message
 
 
 @click.command()
-@click.option('-m', '--input_method', 'input_method', default='defaults', help='Method by which ' +
+@click.option('-i', '--input_method', 'input_method', default='defaults', help='Method by which ' +
               'to create the YAML configuration file. Valid choices: \'defaults\', \'cli\'.')
 @click.option('-s', '--suite', 'suite', default='hofx', help='If using defaults for input_method ' +
               'this option is used to determine which suite to obtain the defaults for.')
@@ -29,7 +29,9 @@ from swell.utilities.welcome_message import write_welcome_message
               'platform specific defaults.')
 @click.option('-o', '--override', 'override', default=None, help='After generating the config ' +
               'file parameters inside can be overridden using value from the override config file.')
-def main(input_method, suite, platform, override):
+@click.option('-m', '--models', 'models', multiple=True, default=['geos_atmosphere'], help='Enter' + 
+              'model components.')
+def main(input_method, suite, platform, override, models):
 
     # Welcome message
     # ---------------
@@ -37,7 +39,7 @@ def main(input_method, suite, platform, override):
 
     # Create suites object
     # --------------------
-    config_file = prepare_config(input_method, suite, platform, override)
+    config_file = prepare_config(input_method, suite, platform, override, models)
 
 
 # --------------------------------------------------------------------------------------------------
