+++ src/swell/deployment/prep_config_cli.py
@@ -14,6 +14,7 @@ import re
 import sys
 
 import questionary
+from questionary import Choice
 
 from swell.deployment.prep_config_base import PrepConfigBase
 from swell.swell_path import get_swell_path
@@ -23,125 +24,18 @@ from swell.swell_path import get_swell_path
 
 class PrepConfigCli(PrepConfigBase):
 
-    def execute(self, dictionary=None):
-
-        # Set dictionary to use in this scope
-        if dictionary is None:
-            dictionary = self.dictionary
-
-        print(f"Now editing the {self.directory.split('/')[-1]} YAML file.")
-
-        # Get final key in yaml before fixed options
-        self.dictionary = self.open_dictionary()
-        key_list = list(self.dictionary.keys())
-        if 'fixed_options' in key_list:
-            fixed_idx = key_list.index('fixed_options')
-            key_list.pop(fixed_idx)
-        end_key = key_list[-1]
-
-        self.key_list = key_list
-
-        for key in dictionary:
-            # Element dictionary
-            el_dict = dictionary[key]
-
-            if key != 'fixed_options':
-
-                # Validate the element dictionary
-                self.validate_dictionary(el_dict)
-
-                # Extract type
-                type = el_dict['type']
-
-                # If the type is not another file add to dictionary
-                if 'file' not in type:
-
-                    # Check that the key does not have a dependency
-                    depends_flag = True
-                    if 'depends' in el_dict.keys():
-                        dep_key = el_dict['depends']['key']
-                        dep_val = el_dict['depends']['value']
-                        if self.experiment_dict[dep_key] != dep_val:
-                            depends_flag = False
-
-                    # In this case the key is not expected to refer to a sub dictionary but have
-                    # everything needed in the elements dictionary
-                    if depends_flag:
-                        el_dict['default_value'] = self.check_widgets(key, el_dict)
-                        if key == end_key:
-                            change_check = self.before_next()
-                            if isinstance(change_check, dict):
-                                el_dict = change_check
-                        self.add_to_experiment_dictionary(key, el_dict)
-
-                elif 'file-drop-list' in type:
-
-                    # Add the choice to the dictionary
-                    # If you wanted more suite options, you'd need to add directories for them at
-                    # the suites/ level
-                    el_dict['default_value'] = self.check_widgets(key, el_dict)
-                    change_check = self.before_next()
-                    if isinstance(change_check, dict):
-                        el_dict = change_check
-                    self.add_to_experiment_dictionary(key, el_dict)
-
-                    # In this case the key refers to a single sub dictionary that involves opening
-                    # that dictionary and recursively calling this routine.
-
-                    # First append the directory and filename to denote moving to the sub dictionary
-                    self.append_directory_and_filename(el_dict['default_value'])
-
-                    # Open next level down dictionary and recursively add
-                    self.execute(self.open_dictionary())
-
-                    # As we come back from the sub dictionary subtract the directory and filename
-                    self.subtract_directory_and_filename()
-
-                elif 'file-check-list' in type:
-
-                    # Add the choice to the dictionary
-                    el_dict['default_value'] = self.check_widgets(key, el_dict)
-                    change_check = self.before_next()
-                    if isinstance(change_check, dict):
-                        el_dict = change_check
-                    self.add_to_experiment_dictionary(key, el_dict)
-
-                    # In this case the key asks the user to provide a list of items that correspond
-                    # to sub dictionaries. Inside a loop this method is called recursively.
-                    options = el_dict['default_value']
-                    for option in options:
-
-                        # If the key is models change the internal model to match this model
-                        if key == 'model_components':
-                            self.update_model(option)
-
-                        self.append_directory_and_filename(option)
-                        self.execute(self.open_dictionary())
-                        self.subtract_directory_and_filename()
-
-                        if key == 'model_components':
-                            self.update_model(None)
-
-            else:
-
-                for fixed_key in el_dict:
-
-                    self.add_to_experiment_dictionary(fixed_key, el_dict[fixed_key])
-
-        return
-
-    def check_widgets(self, key, val):
-        print('\n')
+    def get_answer(self, key, val):
+        # Set questionary variable
         widget_type = val['type']
         quest = val['prompt']
         default = val['default_value']
+
+        # Check questionary input type and ask user
+        print('\n')
         if widget_type == 'string':
             answer = self.make_string_widget(quest, default, questionary.text)
         elif 'drop-list' in widget_type:
-            if 'file' in widget_type:
-                options = 'file'
-            elif 'string' in widget_type:
-                options = val['options']
+            options = val['options']
             answer = self.make_drop_widget(key, quest, options, default, questionary.select)
         elif widget_type == 'boolean':
             answer = self.make_boolean(quest, default, questionary.confirm)
@@ -150,10 +44,7 @@ class PrepConfigCli(PrepConfigBase):
         elif widget_type == 'iso-duration':
             answer = self.make_duration(quest, default, questionary.text)
         elif 'check-list' in widget_type:
-            if 'file' in widget_type:
-                options = 'file'
-            elif 'string' in widget_type:
-                options = val['options']
+            options = val['options']
             answer = self.make_check_widget(quest, options, default, questionary.checkbox)
         else:
             answer = default
@@ -167,43 +58,73 @@ class PrepConfigCli(PrepConfigBase):
 
         return answer
 
+    # ----------------------------------------------------------------------------------------------
+
+    def get_models(self):
+
+        model_options = glob.glob(os.path.join(self.model_path, '*'))
+        model_options.sort()
+
+        print('TEST', model_options)
+
+        model_options = [os.path.basename(x) for x in model_options]
+
+        choices = []
+
+        for mod in model_options:
+            if mod in self.model_choices:
+                choices.append(Choice(mod, checked=True))
+            else:
+                choices.append(Choice(mod, checked=False))
+
+        selected_models = self.make_check_widget('Which models?', choices, default = None, prompt=questionary.checkbox)
+
+        if 'None' in selected_models:
+            selected_models = []
+
+        return selected_models
+
+    # ----------------------------------------------------------------------------------------------
+
     def make_string_widget(self, quest, default, prompt):
         answer = prompt(f"{quest} [{default}]", default=default).ask()
 
         return answer
 
+    # ----------------------------------------------------------------------------------------------
+
     def make_drop_widget(self, method, quest, options, default, prompt):
-        if options == 'file':
-            new_path = os.path.join(self.directory, '*/')
-            suite_list = [x.split('/')[-2] for x in glob.glob(new_path)]
+        if options == 'use_method':
+            if method == 'platform':
+                method_dir = 'deployment/platforms/'
+            elif method == 'suite_to_run':
+                method_dir = 'suites_new/suite_to_run'
+            new_path = os.path.join(os.path.dirname(self.directory), method_dir, '*')
+            suite_list = [os.path.splitext(os.path.basename(x))[0] for x in glob.glob(new_path)]
             # Make sure no python directories are included
-            suite_list = list(filter(lambda a: a != '__pycache__', suite_list))
+            suite_list = list(filter(lambda a: a != '__pycache__' and a != '__init__', suite_list))
             choices = suite_list
         else:
-            if options == 'use_method':
-                if 'platform' in method:
-                    method_dir = 'deployment/platforms/'
-                new_path = os.path.join(os.path.dirname(self.directory), method_dir, '*/')
-                suite_list = [x.split('/')[-2] for x in glob.glob(new_path)]
-                # Make sure no python directories are included
-                suite_list = list(filter(lambda a: a != '__pycache__', suite_list))
-                options = suite_list
             choices = options
         answer = prompt(quest, choices=choices, default=default).ask()
 
         return answer
+
+    # ----------------------------------------------------------------------------------------------
 
     def make_boolean(self, quest, default, prompt):
         answer = prompt(quest, default=default, auto_enter=False).ask()
 
         return answer
 
+    # ----------------------------------------------------------------------------------------------
+
     def make_datetime(self, quest, default, prompt):
 
         class dtValidator(questionary.Validator):
             def validate(self, document):
                 r = re.compile('\d\d\d\d-\d\d-\d\dT\d\d:\d\d:\d\dZ')  # noqa
-                if r.match(document.text) is None:
+                if r.match(document.text) is None and document.text != 'EXIT':
                     raise questionary.ValidationError(
                         message="Please enter a datetime with the following format: " +
                                 "YYYY-MM-DDThh:mm:ssZ",
@@ -215,12 +136,14 @@ class PrepConfigCli(PrepConfigBase):
 
         return answer
 
+    # ----------------------------------------------------------------------------------------------
+
     def make_duration(self, quest, default, prompt):
 
         class durValidator(questionary.Validator):
             def validate(self, document):
                 r = re.compile('PT\d{1,2}H')  # noqa
-                if r.match(document.text) is None:
+                if r.match(document.text) is None and document.text != 'EXIT':
                     raise questionary.ValidationError(
                         message="Please enter a duration with the following format: PThhH",
                         cursor_position=len(document.text),
@@ -246,54 +169,53 @@ class PrepConfigCli(PrepConfigBase):
 
         return answer
 
+    # ----------------------------------------------------------------------------------------------
+
     def make_check_widget(self, quest, options, default, prompt):
-        if options == 'file':
-            dir_list = os.listdir(self.directory)
-            new_path = os.path.join(self.directory, '*/')
-            suite_list = [x.split('/')[-2] for x in glob.glob(new_path)]
-            choices = suite_list
-            default = None
+        if options == 'use_method':
+            choices = default
+            default = default[0]
+            if self.model is not None:
+                files = glob.glob(os.path.join(self.install_path,
+                                               'configuration/jedi/interfaces',
+                                               self.model,
+                                               'observations/*.yaml'))
+                # Do not include obsop_name_map.yaml in the list of observations
+                files = list(filter(lambda a: 'obsop_name_map' not in a, files))
+                choices = [os.path.splitext(os.path.basename(x))[0] for x in files]
         else:
-            if options == 'use_method':
-                choices = default
-                default = default[0]
-                if self.model is not None:
-                    files = glob.glob(os.path.join(self.install_path,
-                                                   'configuration/jedi/interfaces',
-                                                   self.model,
-                                                   'observations/*.yaml'))
-                    # Do not include obsop_name_map.yaml in the list of observations
-                    files = list(filter(lambda a: 'obsop_name_map' not in a, files))
-                    choices = [x.split('/')[-1].split('.')[0] for x in files]
-            else:
-                choices = options
-                if default is None:
-                    pass
-                else:
-                    default = default[0]
-        answer = prompt(quest, choices=choices,
-                        default=default,
+            choices = options
+
+        answer = prompt(quest, choices=choices, default=default,
                         validate=lambda text: True if text != []
                         else 'Please select one option').ask()
         return answer
+
+    # ----------------------------------------------------------------------------------------------
 
     def before_next(self):
         changer = self.make_boolean('Do you wish to change any of your entries?',
                                     False,
                                     questionary.confirm)
         if changer:
-            keys = self.key_list
+            keys = self.exec_keys
+            for k in keys:
+                if k not in list(self.dictionary.keys()):
+                    non_exec_idx = keys.index(k)
+                    keys.pop(non_exec_idx)
+            # Show user key change options and retrieve new values
             change_keys = self.make_check_widget('Which elements would you like to change?',
                                                  keys,
                                                  None,
                                                  questionary.checkbox)
+
             for k in change_keys:
                 changed_dict = self.dictionary[k]
                 new_default_value = self.check_widgets(k, changed_dict)
                 if 'file' in changed_dict['type']:
                     changed_dict['default_value'] = new_default_value
                     return changed_dict
-                elif k == self.key_list[-1]:
+                elif k == self.exec_keys[-1]:
                     changed_dict['default_value'] = new_default_value
                     return changed_dict
                 else:
